# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ristocats API",
    "description": "This is the API for the project RistoCats, a restaurant management system.",
    "contact": {
      "url": "https://github.com/OldKingAllant/RistoCats"
    },
    "license": {
      "name": "GNU GPL 3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "version": "1.0"
  },
...
  "paths": {
...
    "/api/users/verify": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Verify jwt token",
        "description": "Verifies jwt token and returns 200 or 401/400",
        "operationId": "verifyToken",
        "parameters": [
          {
            "name": "jwt",
            "in": "query",
            "description": "jwt token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidRequest"
                }
              }
            }
          },
          "400": {
            "description": "query is empty"
          },
          "401": {
            "description": "invalid/missing token"
          }
        }
      }
    },
...
    "/api/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Perform login",
        "description": "Generates JWT and returns it",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "jwt created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jwt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "empty query"
          },
          "401": {
            "description": "invalid google code/invalid email/..."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ValidRequest": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "example": "true"
          }
        }
      },
...
    }
  }
}